+++ src/main/java/com/alvinskylers/products/dto/ProductCategoryResponse.java
@@ -2,11 +2,9 @@ package com.alvinskylers.products.dto;
 
 import lombok.Builder;
 
-import java.math.BigInteger;
-
 @Builder
 public record ProductCategoryResponse(
-        BigInteger id,
+        Long id,
         String name,
         String description
 ) {
