+++ src/main/java/com/alvinskylers/products/exception/ProductCategoryNotFoundException.java
@@ -3,7 +3,7 @@ package com.alvinskylers.products.exception;
 import java.math.BigInteger;
 
 public class ProductCategoryNotFoundException extends ResourceNotFoundException {
-    public ProductCategoryNotFoundException(BigInteger id) {
+    public ProductCategoryNotFoundException(Long id) {
         super("product category with id: " + id + " was not found.");
     }
 }
